requirements.txt
discord.py>=1.6.0
discord-py-slash-command>=1.1.0 
peewee>=3.14.0
mcstatus>=5.1.2
aioredis>=1.3.1

